Ajustado controller de pacientes.

+++ Projeto1_IF/Controllers/TbPacientesController.cs
@@ -2,30 +2,52 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Projeto1_IF.Data;
 using Projeto1_IF.Models;
 
 namespace Projeto1_IF.Controllers
 {
+    [Authorize]
     public class TbPacientesController : Controller
     {
         private readonly db_IFContext _context;
+        private readonly ApplicationDbContext _identity;
+        private readonly UserManager<IdentityUser> _userManager;
 
-        public TbPacientesController(db_IFContext context)
+        public TbPacientesController(
+            db_IFContext context,
+            UserManager<IdentityUser> userManager,
+            ApplicationDbContext identity
+        )
         {
             _context = context;
+            _userManager = userManager;
+            _identity = identity;
         }
 
-        // GET: TbPacientes
+        [Authorize(Roles = "Medico, Nutricionista")]
         public async Task<IActionResult> Index()
         {
-            var db_IFContext = _context.TbPaciente.Include(t => t.IdCidadeNavigation);
+            var idProfissional = await GetProfissionalIdByUserIdAsync();
+
+            if (idProfissional == null)
+            {
+                throw new InvalidOperationException("Não foi possível obter o ID do profissional associado ao usuário atual.");
+            }
+
+            var db_IFContext = _context.TbPaciente.AsQueryable()
+                .AsNoTracking()
+                .Where(p => p.TbMedicoPaciente.Any(mp => mp.IdProfissional == idProfissional));
+
             return View(await db_IFContext.ToListAsync());
         }
 
-        // GET: TbPacientes/Details/5
+        [Authorize(Roles = "Medico, Nutricionista")]
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -33,10 +55,19 @@ namespace Projeto1_IF.Controllers
                 return NotFound();
             }
 
+            var idProfissional = await GetProfissionalIdByUserIdAsync();
+
+            if (idProfissional == null)
+            {
+                throw new InvalidOperationException("Não foi possível obter o ID do profissional associado ao usuário atual.");
+            }
+
             var tbPaciente = await _context.TbPaciente
                 .Include(t => t.IdCidadeNavigation)
                 .ThenInclude(s => s.IdEstadoNavigation)
-                .FirstOrDefaultAsync(m => m.IdPaciente == id);
+                .AsNoTracking()
+                .FirstOrDefaultAsync(m => m.IdPaciente == id && m.TbMedicoPaciente.Any(mp => mp.IdProfissional == idProfissional));
+
             if (tbPaciente == null)
             {
                 return NotFound();
@@ -45,18 +76,16 @@ namespace Projeto1_IF.Controllers
             return View(tbPaciente);
         }
 
-        // GET: TbPacientes/Create
+        [Authorize(Roles = "Medico, Nutricionista")]
         public IActionResult Create()
         {
             ViewData["IdCidade"] = new SelectList(_context.TbCidade, "IdCidade", "Nome");
             return View();
         }
 
-        // POST: TbPacientes/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Medico, Nutricionista")]
         public async Task<IActionResult> Create([Bind("IdPaciente,Nome,Rg,Cpf,DataNascimento,NomeResponsavel,Sexo,Etnia,Endereco,Bairro,IdCidade,TelResidencial,TelComercial,TelCelular,Profissao,FlgAtleta,FlgGestante")] TbPaciente tbPaciente)
         {
             try
@@ -64,6 +93,25 @@ namespace Projeto1_IF.Controllers
                 if (ModelState.IsValid)
                 {
                     _context.Add(tbPaciente);
+                    await _context.SaveChangesAsync();
+
+                    var idProfissional = await GetProfissionalIdByUserIdAsync();
+
+                    if (idProfissional != null)
+                    {
+                        TbMedicoPaciente medicoPaciente = new TbMedicoPaciente
+                        {
+                            IdProfissional = (int) idProfissional,
+                            IdPaciente = tbPaciente.IdPaciente
+                        };
+
+                        _context.Add(medicoPaciente);
+                    }
+                    else
+                    {
+                        throw new InvalidOperationException("Não foi possível obter o ID do profissional associado ao usuário atual.");
+                    }
+
                     await _context.SaveChangesAsync();
                     return RedirectToAction(nameof(Index));
                 }
@@ -81,7 +129,7 @@ namespace Projeto1_IF.Controllers
             return View(tbPaciente);
         }
 
-        // GET: TbPacientes/Edit/5
+        [Authorize(Roles = "Medico, Nutricionista")]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -89,7 +137,19 @@ namespace Projeto1_IF.Controllers
                 return RedirectToAction("Erro", "Home");
             }
 
-            var tbPaciente = await _context.TbPaciente.Include(t => t.IdCidadeNavigation).ThenInclude(t => t.IdEstadoNavigation).FirstOrDefaultAsync(s => s.IdPaciente == id);
+            var idProfissional = await GetProfissionalIdByUserIdAsync();
+
+            if (idProfissional == null)
+            {
+                throw new InvalidOperationException("Não foi possível obter o ID do profissional associado ao usuário atual.");
+            }
+
+            var tbPaciente = await _context.TbPaciente
+                .Include(t => t.IdCidadeNavigation)
+                .ThenInclude(s => s.IdEstadoNavigation)
+                .AsNoTracking()
+                .FirstOrDefaultAsync(m => m.IdPaciente == id && m.TbMedicoPaciente.Any(mp => mp.IdProfissional == idProfissional));
+
             if (tbPaciente == null)
             {
                 return NotFound();
@@ -99,11 +159,9 @@ namespace Projeto1_IF.Controllers
             return View(tbPaciente);
         }
 
-        // POST: TbPacientes/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Medico, Nutricionista")]
         public async Task<IActionResult> EditPost(int? id)
         {
             if (id == null)
@@ -111,7 +169,19 @@ namespace Projeto1_IF.Controllers
                 return RedirectToAction("Erro", "Home");
             }
 
-            var tbPaciente = await _context.TbPaciente.Include(t => t.IdCidadeNavigation).ThenInclude(t => t.IdEstadoNavigation).FirstOrDefaultAsync(s => s.IdPaciente == id);
+            var idProfissional = await GetProfissionalIdByUserIdAsync();
+
+            if (idProfissional == null)
+            {
+                throw new InvalidOperationException("Não foi possível obter o ID do profissional associado ao usuário atual.");
+            }
+
+            var tbPaciente = await _context.TbPaciente
+                .Include(t => t.IdCidadeNavigation)
+                .ThenInclude(s => s.IdEstadoNavigation)
+                .AsNoTracking()
+                .FirstOrDefaultAsync(m => m.IdPaciente == id && m.TbMedicoPaciente.Any(mp => mp.IdProfissional == idProfissional));
+
             if (tbPaciente == null)
             {
                 return NotFound();
@@ -120,16 +190,30 @@ namespace Projeto1_IF.Controllers
             if (await TryUpdateModelAsync<TbPaciente>(
                 tbPaciente,
                 "",
-                s => s.IdPaciente, s => s.Nome, s => s.Rg, s => s.Cpf, s => s.DataNascimento, s => s.NomeResponsavel,
-                s => s.Sexo, s => s.Etnia, s => s.Endereco, s => s.Bairro, s => s.IdCidade, s => s.TelResidencial,
-                s => s.TelComercial, s => s.TelCelular, s => s.Profissao, s => s.FlgAtleta, s => s.FlgGestante))
+                s => s.IdPaciente,
+                s => s.Nome,
+                s => s.Rg,
+                s => s.Cpf,
+                s => s.DataNascimento,
+                s => s.NomeResponsavel,
+                s => s.Sexo,
+                s => s.Etnia,
+                s => s.Endereco,
+                s => s.Bairro,
+                s => s.IdCidade,
+                s => s.TelResidencial,
+                s => s.TelComercial,
+                s => s.TelCelular,
+                s => s.Profissao,
+                s => s.FlgAtleta,
+                s => s.FlgGestante))
             {
                 try
                 {
                     await _context.SaveChangesAsync();
                     return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateException /* ex */)
+                catch (DbUpdateException)
                 {
                     ModelState.AddModelError("", "Unable to save changes. " +
                         "Try again, and if the problem persists, " +
@@ -142,7 +226,7 @@ namespace Projeto1_IF.Controllers
             return View(tbPaciente);
         }
 
-        // GET: TbPacientes/Delete/5
+        [Authorize(Roles = "Medico, Nutricionista")]
         public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
@@ -150,11 +234,18 @@ namespace Projeto1_IF.Controllers
                 return NotFound();
             }
 
+            var idProfissional = await GetProfissionalIdByUserIdAsync();
+
+            if (idProfissional == null)
+            {
+                throw new InvalidOperationException("Não foi possível obter o ID do profissional associado ao usuário atual.");
+            }
+
             var tbPaciente = await _context.TbPaciente
                 .Include(t => t.IdCidadeNavigation)
                 .ThenInclude(s => s.IdEstadoNavigation)
                 .AsNoTracking()
-                .FirstOrDefaultAsync(m => m.IdPaciente == id);
+                .FirstOrDefaultAsync(m => m.IdPaciente == id && m.TbMedicoPaciente.Any(mp => mp.IdProfissional == idProfissional));
 
             if (tbPaciente == null)
             {
@@ -171,33 +262,73 @@ namespace Projeto1_IF.Controllers
             return View(tbPaciente);
         }
 
-        // POST: TbPacientes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Medico, Nutricionista")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var tbPaciente = await _context.TbPaciente.FindAsync(id);
+            var idProfissional = await GetProfissionalIdByUserIdAsync();
+
+            if (idProfissional == null)
+            {
+                throw new InvalidOperationException("Não foi possível obter o ID do profissional associado ao usuário atual.");
+            }
+
+            var tbPaciente = await _context.TbPaciente
+                .Include(t => t.IdCidadeNavigation)
+                .ThenInclude(s => s.IdEstadoNavigation)
+                .AsNoTracking()
+                .FirstOrDefaultAsync(m => m.IdPaciente == id && m.TbMedicoPaciente.Any(mp => mp.IdProfissional == idProfissional));
+
             if (tbPaciente == null)
             {
-                return RedirectToAction(nameof(Index));
+                return NotFound();
             }
 
             try
             {
+                var tbMedicoPaciente = await _context.TbMedicoPaciente
+                    .FirstOrDefaultAsync(m => m.IdPaciente == id && m.IdProfissional == idProfissional);
+
+                if (tbMedicoPaciente == null) 
+                { 
+                    return NotFound(); 
+                }
+
+                _context.TbMedicoPaciente.Remove(tbMedicoPaciente);
+                await _context.SaveChangesAsync();
+
                 _context.TbPaciente.Remove(tbPaciente);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            catch (DbUpdateException /* ex */)
+            catch (DbUpdateException)
             {
-                //Log the error (uncomment ex variable name and write a log.)
                 return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
             }
         }
 
+
         private bool TbPacienteExists(int id)
         {
             return _context.TbPaciente.Any(e => e.IdPaciente == id);
+        }
+
+        private async Task<int?> GetProfissionalIdByUserIdAsync()
+        {
+            var usuario = await _userManager.GetUserAsync(User);
+
+            if (usuario == null)
+            {
+                return null;
+            }
+
+            var profissionalId = await _context.TbProfissional
+                .Where(p => p.IdUser == usuario.Id)
+                .Select(p => p.IdProfissional)
+                .FirstOrDefaultAsync();
+
+            return profissionalId;
         }
     }
 }
